separate readings tests into own files
