+++ src/server/test/web/readingsBarMeterBasic.js
@@ -1,0 +1,43 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
+
+/*
+    This file tests the readings retrieval API for basic function for bar charts for meters.
+    See: https://github.com/OpenEnergyDashboard/DesignDocs/blob/main/testing/testing.md for information.
+*/
+
+const { chai, mocha, expect, app } = require('../common');
+const { prepareTest,
+    getUnitId,
+    ETERNITY,
+    METER_ID,
+    HTTP_CODE,
+    unitDatakWh,
+    conversionDatakWh,
+    meterDatakWh } = require('../../util/readingsUtils');
+
+mocha.describe('readings API', () => {
+    mocha.describe('readings test, test if data returned by API is as expected', () => {
+        mocha.describe('for bar charts', () => {
+            mocha.describe('basic for meters', () => {
+                mocha.it('response should have a valid reading, startTimestamp, and endTimestamp', async () => {
+                    await prepareTest(unitDatakWh, conversionDatakWh, meterDatakWh);
+                    // Get the unit ID since the DB could use any value.
+                    const unitId = await getUnitId('kWh');
+                    const res = await chai.request(app).get(`/api/unitReadings/bar/meters/${METER_ID}`)
+                        .query({
+                            timeInterval: ETERNITY.toString(),
+                            barWidthDays: 1,
+                            graphicUnitId: unitId
+                        });
+                    expect(res).to.be.json;
+                    expect(res).to.have.status(HTTP_CODE.OK);
+                    expect(res.body).to.have.property(`${METER_ID}`).to.have.property('0').to.have.property('reading');
+                    expect(res.body).to.have.property(`${METER_ID}`).to.have.property('0').to.have.property('startTimestamp');
+                    expect(res.body).to.have.property(`${METER_ID}`).to.have.property('0').to.have.property('endTimestamp');
+                });
+            });
+        });
+    });
+});
